feat: migrate landing page to MUI
Bold-ui removed
Favicons updated
Roboto set as app font
