+++ packages/dashboard/package.json
@@ -29,6 +29,7 @@
     "@apollo/client": "^3.3.11",
     "@emotion/react": "^11.4.1",
     "@emotion/styled": "^11.3.0",
+    "@fontsource/roboto": "^4.5.1",
     "@mui/icons-material": "^5.0.5",
     "@mui/lab": "5.0.0-alpha.53",
     "@mui/material": "^5.0.6",
@@ -36,7 +37,6 @@
     "@mui/x-data-grid": "^5.0.0-beta.6",
     "@popperjs/core": "^2.4.0",
     "@sorry-cypress/common": "1.0.0",
-    "bold-ui": "^1.0.0-beta.40",
     "date-fns": "^2.21.3",
     "dotenv": "^8.2.0",
     "ejs": "^2.7.1",
@@ -90,7 +90,9 @@
     "npm-run-all": "^4.1.5",
     "prettier": "^2.1.2",
     "style-loader": "^1.0.0",
+    "svg-url-loader": "^7.1.1",
     "typescript": "^4.1.5",
+    "url-loader": "^4.1.1",
     "webpack": "^4.41.5",
     "webpack-cli": "^3.3.3",
     "webpack-dev-middleware": "^3.7.2",
