+++ packages/dashboard/src/components/layout/layout.tsx
@@ -1,7 +1,6 @@
 import { Box, Container, Toolbar, useMediaQuery } from '@mui/material';
 import { useTheme } from '@mui/material/styles';
 import { useLocalStorage } from '@sorry-cypress/dashboard/hooks';
-import { WithMaterial } from '@sorry-cypress/dashboard/lib/material';
 import React, { PropsWithChildren } from 'react';
 import { Header } from './header';
 import { Sidebar } from './sidebar';
@@ -23,10 +22,8 @@ export const Layout = ({ children }: PropsWithChildren<any>) => {
     <Box
       sx={{ display: 'flex', minHeight: '100vh', backgroundColor: '#E8E8EC' }}
     >
-      <WithMaterial>
-        <Sidebar open={open} onToggleSidebar={toggleSidebar} />
-        <Header open={open} onMenuClick={toggleSidebar} />
-      </WithMaterial>
+      <Sidebar open={open} onToggleSidebar={toggleSidebar} />
+      <Header open={open} onMenuClick={toggleSidebar} />
       <Container
         maxWidth="xl"
         component="main"
@@ -37,9 +34,7 @@ export const Layout = ({ children }: PropsWithChildren<any>) => {
           width: 'auto',
         }}
       >
-        <WithMaterial>
-          <Toolbar />
-        </WithMaterial>
+        <Toolbar />
         {children}
       </Container>
     </Box>
