+++ packages/dashboard/src/components/ui/index.ts
@@ -1,11 +1,7 @@
 export * from './card';
-export * from './centered';
 export * from './chip';
 export * from './expandableArea';
-export * from './flexRow';
-export * from './headerLink';
 export * from './pad';
-export * from './pageControls';
 export * from './paper';
 export * from './searchField';
 export * from './toolbar';
