--- packages/dashboard/src/components/ui/pageControls.tsx
@@ -1,8 +1,0 @@
-import React, { PropsWithChildren } from 'react';
-import { FlexRow } from './flexRow';
-
-const PageControls = ({ children }: PropsWithChildren<unknown>) => (
-  <FlexRow>{children}</FlexRow>
-);
-
-export { PageControls };
