+++ packages/dashboard/src/components/ui/searchField.tsx
@@ -1,7 +1,6 @@
 import { Search } from '@mui/icons-material';
 import { InputAdornment, TextField, Theme } from '@mui/material';
 import useDebounce from '@sorry-cypress/dashboard/hooks/useDebounce';
-import { WithMaterial } from '@sorry-cypress/dashboard/lib/material';
 import React, { ChangeEvent, useState } from 'react';
 
 export type OnSearch = (value: string) => unknown;
@@ -26,38 +25,36 @@ const SearchField = ({ placeholder, onSearch, disabled }: SearchFieldProps) => {
   };
 
   return (
-    <WithMaterial>
-      <TextField
-        fullWidth
-        size="small"
-        InputProps={{
-          startAdornment: (
-            <InputAdornment position="start">
-              <Search />
-            </InputAdornment>
-          ),
-        }}
-        inputProps={{
-          sx: {
-            pt: 0.8,
-            pb: 0.8,
-            width: { lg: 180 },
-            transition: (theme: Theme) =>
-              theme.transitions.create('width', {
-                duration: theme.transitions.duration.short,
-              }),
-            '&:focus': {
-              width: { lg: 300 },
-            },
+    <TextField
+      fullWidth
+      size="small"
+      InputProps={{
+        startAdornment: (
+          <InputAdornment position="start">
+            <Search />
+          </InputAdornment>
+        ),
+      }}
+      inputProps={{
+        sx: {
+          pt: 0.8,
+          pb: 0.8,
+          width: { lg: 180 },
+          transition: (theme: Theme) =>
+            theme.transitions.create('width', {
+              duration: theme.transitions.duration.short,
+            }),
+          '&:focus': {
+            width: { lg: 300 },
           },
-        }}
-        variant="outlined"
-        placeholder={placeholder}
-        value={value}
-        onChange={handleOnChange}
-        disabled={disabled}
-      />
-    </WithMaterial>
+        },
+      }}
+      variant="outlined"
+      placeholder={placeholder}
+      value={value}
+      onChange={handleOnChange}
+      disabled={disabled}
+    />
   );
 };
