+++ packages/dashboard/src/index.ejs
@@ -15,7 +15,6 @@
     <meta name="theme-color" content="#ffffff" />
 
     <title>Sorry-cypress dashboard</title>
-    <link href="/fonts.css" rel="stylesheet" />
 
     <script>
       window.__sorryCypressEnvironment = <?- SORRY_CYPRESS_ENVIRONMENT ?>;
