+++ packages/dashboard/src/resources/custom.d.ts
@@ -1,0 +1,4 @@
+declare module '@sorry-cypress/dashboard/resources/*.svg' {
+  const content: string;
+  export default content;
+}
